Feature/equidistant timeseries (#21)
Implement a new feature to interpolate `TimeseriesDataset`s.
The resulting `TimeseriesDataset`s will have equidistant points in the half-open interval between the start and end point of the dataset (i.e. the period for which data was requested initially). However, these points might still have `NaN` values depending on the interpolation method. Default method is forward-fill with no limit on the number of points to fill.