Improve client handling and refactor OAuth client
OAuth sessions are used more than once now, removing a lot of redundant calls to obtain a new token.
A new token is created when the session is about to expire or if there is a scope mismatch.
Using sessions inherently reaps the benefits of underlying connection pooling mechanisms.

The OAuth client has been refactored along the way.