Create pull request template with checklist (#78)
As discussed, each PR should now have a checklist so that important non-automatable checks are not forgotten.

I have added things that might not be obvious such as, e.g., "unit tests". If code is changed that is already covered by unit tests, the coverage might still be fine but there might also be new aspects of the new code that should be tested for additionally. So it is better to think twice.