AC: retrieving functional locations from asset central (#86)
Functional Locations can be used in almost the same way as Equipment in PAI. For some customer scenarios it is required to work with Functional Locations in the same way as with Equipment. Hence sailor should provide fetching the relevant information.

This feature covers the basic functionality for retrieving functional locations with the filter language used in the other methods. More advanced functionality, similar to the equipment model could be implemented further.