Analysis tables for systems with hierarchical structure (#105)
Supports analysis tables for a system set consisting of systems with a hierarchical structure.
Users can provide a selection dictionary that defines which pieces of equipment and indicators shall be taken into account. If no selection dictionary is provided, the pieces of equipment and indicators are selected that exist for all systems of the provided system set.