Pagination of results for Assetcentral/PAI APIs
Assetcentral/PAI API limits the result set by default which may result in truncated result sets when using Sailor's find_* functions.
The only way to work around this for the end-user is to manually split the expected result into multiple find_* function calls based on some carefully chosen attribute(s) and values.

Each API supports pagination for most entities.
This commit implements pagination and makes use of it for all supported entities. That means that a single find_* function call will now return the entire result set by using fetch pagination in the background.

The focus of this implementation is on completeness and correctness. There are no performance gains to be expected. The feature is implemented using synchronous calls.